Adding custom validator for Note entity
Adding the custom validator to comply with a DB constraint in the notes table.
The note must be associated with at least one category, product, or supplier.
`CONSTRAINT notes_at_least_one_id_required CHECK (category_id IS NOT NULL OR supplier_id IS NOT NULL OR product_id IS NOT NULL)`
To do this i'm creating a custom validator annotation @AtLeastOneNotNull to check for this constraint on the hibernate side.